requirements.txt
Pillow==10.0.0
pymongo==4.4.1
pyTelegramBotAPI==4.12.0
python-dotenv==1.0.0